Load HWAS module earlier in ext init tasks
We were seeing crashes when an error occurred early on in the boot.
Errors get processed immediatly following the loading of the IMPI
module, this is because at this point we can send the logs to the BMC.
If the error had a sensor callout it would attempt to branch to
hwas code, prior to this commit, the HWAS commit would not be loaded in
at this point causing an instruction exception. By moving the loading
of the HWAS module to be prior to the loading of the IMPI module we
should avoid this issue.

Change-Id: I012e717577dc717dac67d9cf25a5091bb4a0fd84
CQ: SW455241
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70939
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>