Force VPD IO through HW (Axone) ,add plat function to get OCMB VPD
This commit follows up on previous work that pulled in the code
that can find the correct EFD given a SPD blob of data which has
the DDR SPD and the EFD for the DDIMM. This commit adds ocmb_spd.C
which provides a DeviceFW::SPD read interface for OCBM targets,
and hooks is up to the platform implementation of the FAPI
interface platGetVPD for OCMB target. Also this commit forces all
VPD IO in Axone to go through HW, which will actually read from
the new EEPROM cache in pnor.

RTC: 203718
Change-Id: I270500898c422d4c78daa3b917b1b2e5b049e856
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72165
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>