Add temporary Axone simics workarounds to progress IPL
Currently there is no VRM hooked up to the other side of the AVSbus
in simics. The simics team is working on this but for now we need
to skip the istep that calls setup evid to set voltages. This can
be removed when Simics gets this working. Also for now we are will
skip starting checkstop handling early on in the IPL because the
OCC model is not finished. This also can be changed when the model
starts working.

Change-Id: Ia0df49fedae97acceefe07e3f3c903bbe6aac83d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71097
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>