Fix issues w/ axone updates to simics scripts
Simics teams were hitting a problem trying to boot older simics models
because our scripts relied on $MACHINE being set to CEC_MODEL. This
value gets set by linux so we cannot rely on it. Instead use the old
way to lookup MODEL for older version of simics, and for new way,
(axone and beyond) just force MODEL to be "AXONE" for now. Eventually
simics team needs to provide interface to lookup MODEL. Also in this
commit we need to write "0x5A5A5A5A" to the SBE seeprom images to avoid
updates, update the simics backing build, and change the CLI args for
hb startsimics in axone to no longer use -o and always pass in the path
to the combined SBE image/ HBBL image

Change-Id: Ib477e86b0b8aa1a4d44545c62b16fc2cbed2034b
CQ: SW453806
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71323
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: CALEB SCHLOSSIN <calebs@us.ibm.com>
Reviewed-by: VIJAY F. LOBO <vflobo@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>