Statically allocate work pages for PNOR Resource Provider
Any time the PNOR Resource Provider loads in a page that is ECC
protected it allocates 0x1200 bytes to hold the data while it
is decoded.  Continuously allocating 2 pages from the heap leads
to a couple of potential issues:
- fragmentation of memory due to constant new/delete of pages
- extremely low memory situation could prevent new code from
  paging in

The solution here is to place the working buffer inside the PNOR
RP class itself.  This means we will only do a single allocation
of the memory early on in the boot and the memory is always
available.

Change-Id: I7faabb6be7cc89abdb8a8fb962e0623a2bcd5e99
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71546
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>