Set simics xml to match simics model for OCMB port numbering
There was a bug where the i2c controller information had the wrong
port values for OCBM targets 2-8. Correct value for the port for
these targets is 0. In the current axone simics model OCMB 0 is on
port 1 and OCMBs 1-8 are behind a 8-1 mux on port 0. Port was correct
for OCMB 0 and 1 but incorrect for 2-8.

Change-Id: Id34a812e2f278d0bc90beb44ba26b0fec32d2087
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/71700
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>