Update simics build and pull in Axone binary to pnor generation
Previously we were using some CUMULUS binaries as placeholders until
Axone versions were available. We now use the Axone version of the SBE
and HW ref images. Also for MEMD we will now  just fill it in with
zeroed out ECC (all FF's w/ 0 for ECC bit). No plans to use that MEMD now
but setting it aside for DDIMM config overwrites. When we pulled in the
new SBE image we hit issues with the VPD being used not having the correct
sizes. MVPD for Axone is not complete yet so until that gets resolved we
will disable sbe updates for Axone.

RTC: 197497
Change-Id: I470f44d297179556d7c2eb7b210c91a2cd38f23c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72090
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>