Use shell command to look up CEC_TYPE
Using the native 'env' command in simics causes users to have problems
when reloading checkpoints on pre-axone systems. When you reload
the checkpoint CEC_MODEL is not defined and it gives env an error
which crashes the simics run. Using the shell command allows it
to fail to find the CEC_MODEL variable and is able to continue on.
This is actually reverting back to the original behavior prior
to suggestions from simics team.

Change-Id: Ieefd0d73e0651cf233f94a297f8f4ea2f5f0dbfe
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72161
Reviewed-by: Hieu C. Nguyen <hieu.nguyen@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: CALEB SCHLOSSIN <calebs@us.ibm.com>
Reviewed-by: Michael J. Walsh <micwalsh@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>