Move watchdog timer to prevent timeout
The watchdog used to be set in istepdispatcher.C but due to various
changes to PNOR that would cause a watchdog timeout. This commit moves
the watchdog out of impiext code and into impibase which allows the
watchdog to be setup as soon as the ipmi resource provider comes up.
Then once the code reaches istepdispatcher the watchdog is adjusted as
needed.

Change-Id: Ia5f13800db22cd17b0feabc5257f8f6409f54376
CQ:SW453071
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72269
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>