Add calls to available p9a_mss_eff_config HWPs in istep 7.4
New effective configuration HWPs have been introduced in Axone.
Currently p9a_mss_eff_config and p9a_mss_eff_config_thermal are
no-ops but this commit gets the calls in-place so we are ready.
Eventually we need to pull in the rest of the eff_config HWPs
but there are not even no-ops stubs to pull in yet.

Change-Id: Icd71006d31023caa5ef913a8b81694ae9ea3a5ba
RTC: 195552
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72204
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>