Improve performance of Partial Good logic
The function PARTIAL_GOOD::findRulesForTarget() would call
masterProcChipTargetHandle() to figure out if a rule applied to a
specific target based on the ATTR_MODEL of a system. Since finding the
master proc that way is an expensive operation, findRulesForTarget()
will now use an instance variable initialized once instead of
querying for it repeatedly thereby increasing performance substantially.

Change-Id: I6e9c01daf356fe21d76d920f5d3210717f5094c1
CQ:SW450113
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72620
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>