Moving libnvdimm into a resident module
The nvdimm module provides the back-end of a device driver
interface.  That means that it has a REGISTER_ROUTE call that
will run when the module is loaded.  Right now the module is
loaded/unloaded for several isteps, which means we end up
seeing DEVFW_RC_DOUBLE_REGISTRATION errors.  These are only
informational, but they are still indicating a problem.

This change moves the libnvdimm.so module into the list of
extended initservice tasks that get loaded before we start
the isteps.  This is where all of our other device drivers
live.

I cleaned up a leftover call to libmmio.so that was there
from the same kind of change previously.

Change-Id: I3e2e1bc8e273b46b8c297be0d9d75ff3e660fa00
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72731
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>