Delete complextype fields in attributes if they have no value
The targeting binary build will fail if it sees an attribute with
a complex type that has a field with an empty value tag.  However
if the field isn't present at all, we properly fill in the default
value from our xml metadata files.  This change will delete any
field that is found to have an empty value.

Change-Id: I9d9b2b7870bef067d434f4698efac597043db2bf
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72875
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>