Fix getChildByAffinity call in mmio.C and add more traces
There was a bug in mmio.C where we are trying to find the ocmb targets
associated with a proc. We are looking up UNITs which is wrong because
OCMBs are CHIPs so this commit addresses that. Also this commit elevates
some debug traces to real traces to help with bringup.

Change-Id: I9c11ca4a30c24192955474148cf4e61cca16489c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/72964
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>