Fix issues with how we were handling new .so added for Axone
This commit addresses two problems, one is that we were loading and
unloading expaccess.so in the ocmbcomm tests when it is included
in the extended image. Also we had mmio.so listed as a DEP lib in
istep 12, this caused it to get unloaded after that step. This .so
is also part of the extended image. To remove compile errors I added
these new .so's to the listdeps.pl script.

Change-Id: I83292a53e0021c05468605e62f9b00cefc2af268
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/73182
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>