Tweak page eviction parameters to handle low mem better
We are memory constrained in some configurations during 10.2
because we are forced to pin large chunks of memory to hold a
few data partitions.  In some instances we have run out of free
memory completely.  The fact that we don't hit this all the time
indicates it is purely based on timing and the random nature of
page allocations.

A few changes were made to better handle these kinds of low
memory situations:
- Any time we dip into the kernel's private stash of free pages
we will force memory periodics to run.  This should free up
some pages so that the kernel's stash can remain full.
- The size of the kernel's private page stash is increased from
4 to 10.  We need this larger number due to the increased layers
involved in paging code in/out with secureboot enabled.
- The LRU algorithm was modified to keep pages around longer. A
page will not be evicted under normal circumstances until we go
10 refresh cycles without an access (up from 3).  This should
help balance out the increased number of forced periodics that
could now come into play.

Change-Id: I289d866270db9d70cac275d8c98ab185caaade05
CQ: SW458623
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/73189
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>