Edit cxxtest.start.sh script to call runsim correctly
Simserver has made changes for Axone to account for the new style
to the simics directory. start_simics is not longer a valid way
to start simics in the new directory so we must provide the new
command to start simics via the env var START_SIMICS_CMD and adding
the simics directory to the current path variable.

Change-Id: I8a9f259356a4125e22831585e69840c554261b36
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/74778
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>