Fix bug that was scrambling traces in errorlogs during HBRT
There was a bug in how HB was interpreting the clock time that
PHYP returned in response to a get_clocktime request. The
nsec member of the struct returned was getting the first 32bits
filled in and Hostboot was expecting the value to be in the last
32bits. This change will shift the nsec value down 4 bytes if we
find the back half of the nsec value to be all 0's. I believe the reason
the wrong bits were filled in is because PHYP is returning a packed
struct whose 2nd member is a 32 bit value so it does not get filled in
correctly.

Change-Id: I71f1c2cb672b9979e0b6bdc8b1852544f230b885
CQ: SW448397
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/74823
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>