Stop overwriting OCMB vpd info in standalone and handle DDIMM modules
This commit removes some lines that were overwriting the OCMB's eeproms
with what we thought was "good" VPD. Simics itself now has a better copy
so we will use theirs. When consuming their VPD however we found that
Hostboot was missing support for the DDIMM type modules. This commit
also update the spd driver to handle this new type of memory module.

Change-Id: I97dba2adee4d816f86228d8fbd7aeb671a050e11
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75197
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>