Resize WOF to allow AXONE test pnor to fit inside 64 MB
This commit reduces the size of the WOFDATA section in PNOR from
12 MB -> 6 MB in an effort to get axone pnor test image fitting in
64 MB. We also had to move from using the very larg ZZ WOF image
to use the smaller Zeppelin WOF image. Due to the way the makefile
that generates the pnor image works we always generate "final" versions
of all 3 WOF images (ZZ, Zeppelin, Fleetwood) regardless of what we
are actually building. There were some checks in the genPnorImages
script that would ensure the layout you are using was big enough
for any binary that we generate. I have removed this restriction so
you can now generate binaries that are larger then the space allotted
(we will not use the larger binaries but they still get generated). The
alternative is to edit the makefile to only generate the "final" images
of the sections we want. That is possible but would require a larger
rewrite of the makefile.

Change-Id: Ib52572ad5e3d58accb49f414b6a0c61babf3cefd
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75257
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>