Re-order i2c properties in Axone simics XML to align closer with sim
In the simics model we have 9 OCMBs per processor, 8 behind MC01 that
are behind an 1-to-8 MUX, and 1 behind MC23 that is directly attatched.
Our XML was not portraying this correctly as it was showing that the
directly attached OCMB was OCMB0, which according to our XML is behind
MC01. This commit makes it so OCMBs 0-7 (behind MC01) have i2c attributes
that say they are behind the mux. OCMB8 will be behind MC23 and will be
the directly attached OCMB.

Change-Id: I7ef0cba6021bccc07ca2ccfefa2e02c9ec68eba4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75555
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>