Setup CONFIG_AXONE env variable correctly
In previous attempt to use CONFIG_AXONE env variable as a flag to
decided whether or not to compile certain AXONE only features we
forgot to add in the config variable to HBconfig. Worse is we were
using the wrong env var in the make files as a flag. As a result,
without this change we are not running any Axone HWPs in axone
simics.

Change-Id: I82dd7f86c5ad390a23eab2d2123d1d10ca9edea3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75568
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>